+++ src/Adapter/DbTable/CallbackCheckAdapter.php
@@ -61,7 +61,7 @@ class CallbackCheckAdapter extends AbstractAdapter
      */
     public function setCredentialValidationCallback($validationCallback)
     {
-        if (!is_callable($validationCallback)) {
+        if (! is_callable($validationCallback)) {
             throw new Exception\InvalidArgumentException('Invalid callback provided');
         }
         $this->credentialValidationCallback = $validationCallback;
@@ -96,7 +96,11 @@ class CallbackCheckAdapter extends AbstractAdapter
     protected function authenticateValidateResult($resultIdentity)
     {
         try {
-            $callbackResult = call_user_func($this->credentialValidationCallback, $resultIdentity[$this->credentialColumn], $this->credential);
+            $callbackResult = call_user_func(
+                $this->credentialValidationCallback,
+                $resultIdentity[$this->credentialColumn],
+                $this->credential
+            );
         } catch (\Exception $e) {
             $this->authenticateResultInfo['code']       = AuthenticationResult::FAILURE_UNCATEGORIZED;
             $this->authenticateResultInfo['messages'][] = $e->getMessage();