+++ test/Adapter/Http/ApacheResolverTest.php
@@ -9,13 +9,16 @@
 
 namespace ZendTest\Authentication\Adapter\Http;
 
+use PHPUnit\Framework\TestCase;
 use Zend\Authentication\Adapter\Http\ApacheResolver as Apache;
+use Zend\Authentication\Adapter\Http\Exception\ExceptionInterface;
 
 /**
  * @group      Zend_Auth
  */
-class ApacheResolverTest extends \PHPUnit_Framework_TestCase
+class ApacheResolverTest extends TestCase
 {
+    // @codingStandardsIgnoreStart
     /**
      * Path to test files
      *
@@ -41,6 +44,7 @@ class ApacheResolverTest extends \PHPUnit_Framework_TestCase
      * @var Apache
      */
     protected $_apache;
+    // @codingStandardsIgnoreEnd
 
     /**
      * Sets the paths to files used in this test, and creates a shared resolver instance
@@ -75,7 +79,8 @@ class ApacheResolverTest extends \PHPUnit_Framework_TestCase
      */
     public function testSetFileInvalid()
     {
-        $this->setExpectedException('Zend\\Authentication\\Adapter\\Http\\Exception\\ExceptionInterface', 'Path not readable');
+        $this->expectException(ExceptionInterface::class);
+        $this->expectExceptionMessage('Path not readable');
         $this->_apache->setFile($this->_badPath);
     }
 
@@ -97,8 +102,9 @@ class ApacheResolverTest extends \PHPUnit_Framework_TestCase
      */
     public function testConstructInvalid()
     {
-        $this->setExpectedException('Zend\\Authentication\\Adapter\\Http\\Exception\\ExceptionInterface', 'Path not readable');
-        $apache = new Apache($this->_badPath);
+        $this->expectException(ExceptionInterface::class);
+        $this->expectExceptionMessage('Path not readable');
+        new Apache($this->_badPath);
     }
 
     public function providePasswordFiles()