+++ test/Adapter/Http/AuthTest.php
@@ -7,13 +7,15 @@
 
 namespace ZendTest\Authentication\Adapter\Http;
 
+use PHPUnit\Framework\TestCase;
 use Zend\Authentication\Adapter\Http;
 use Zend\Http\Headers;
 use Zend\Http\Request;
 use Zend\Http\Response;
 
-class AuthTest extends \PHPUnit_Framework_TestCase
+class AuthTest extends TestCase
 {
+    // @codingStandardsIgnoreStart
     /**
      * Path to test files
      *
@@ -55,6 +57,7 @@ class AuthTest extends \PHPUnit_Framework_TestCase
      * @var Http\FileResolver
      */
     protected $_digestResolver;
+    // @codingStandardsIgnoreEnd
 
     /**
      * Set up test configuration
@@ -301,7 +304,8 @@ class AuthTest extends \PHPUnit_Framework_TestCase
         $bad = $this->_digestReply('Bryce', 'ThisIsNotMyPassword');
         $bad = preg_replace(
             '/realm="([^"]+)"/',  // cut out the realm
-            '', $bad
+            '',
+            $bad
         );
 
         $data = $this->_doAuth($bad, 'digest');
@@ -317,7 +321,8 @@ class AuthTest extends \PHPUnit_Framework_TestCase
         $bad = $this->_digestReply('Bryce', 'ThisIsNotMyPassword');
         $bad = preg_replace(
             '/response="([^"]+)"/',  // cut out the realm
-            'response="foobar"', $bad
+            'response="foobar"',
+            $bad
         );
 
         $data = $this->_doAuth($bad, 'both');
@@ -331,8 +336,10 @@ class AuthTest extends \PHPUnit_Framework_TestCase
      * @param  string $scheme       Which authentication scheme to use
      * @return array Containing the result, response headers, and the status
      */
+    // @codingStandardsIgnoreStart
     protected function _doAuth($clientHeader, $scheme)
     {
+        // @codingStandardsIgnoreEnd
         // Set up stub request and response objects
         $request  = new Request;
         $response = new Response;
@@ -382,8 +389,10 @@ class AuthTest extends \PHPUnit_Framework_TestCase
      *
      * @return string
      */
+    // @codingStandardsIgnoreStart
     protected function _digestChallenge()
     {
+        // @codingStandardsIgnoreEnd
         return [
             'type'   => 'Digest ',
             'realm'  => 'realm="' . $this->_digestConfig['realm'] . '"',
@@ -396,8 +405,10 @@ class AuthTest extends \PHPUnit_Framework_TestCase
      *
      * @return string
      */
+    // @codingStandardsIgnoreStart
     protected function _digestReply($user, $pass)
     {
+        // @codingStandardsIgnoreEnd
         $nc       = '00000001';
         $timeout  = ceil(time() / 300) * 300;
         $nonce    = md5($timeout . ':PHPUnit:Zend\Authentication\Adapter\Http');
@@ -427,8 +438,10 @@ class AuthTest extends \PHPUnit_Framework_TestCase
      * @param  string $expected Expected Www-Authenticate header value
      * @return void
      */
+    // @codingStandardsIgnoreStart
     protected function _checkUnauthorized($data, $expected)
     {
+        // @codingStandardsIgnoreEnd
         $result = $status = $headers = null;
         extract($data); // $result, $status, $headers
 
@@ -462,8 +475,10 @@ class AuthTest extends \PHPUnit_Framework_TestCase
      * @param  array $data Authentication results
      * @return void
      */
+    // @codingStandardsIgnoreStart
     protected function _checkOK($data)
     {
+        // @codingStandardsIgnoreEnd
         $result = $status = $headers = null;
         extract($data); // $result, $status, $headers
 
@@ -481,8 +496,10 @@ class AuthTest extends \PHPUnit_Framework_TestCase
      * @param  array $data Authentication results
      * @return void
      */
+    // @codingStandardsIgnoreStart
     protected function _checkBadRequest($data)
     {
+        // @codingStandardsIgnoreEnd
         $result = $status = $headers = null;
         extract($data); // $result, $status, $headers