+++ test/Adapter/Http/FileResolverTest.php
@@ -9,13 +9,15 @@
 
 namespace ZendTest\Authentication\Adapter\Http;
 
+use PHPUnit\Framework\TestCase;
 use Zend\Authentication\Adapter\Http;
 
 /**
  * @group      Zend_Auth
  */
-class FileResolverTest extends \PHPUnit_Framework_TestCase
+class FileResolverTest extends TestCase
 {
+    // @codingStandardsIgnoreStart
     /**
      * Path to test files
      *
@@ -43,6 +45,7 @@ class FileResolverTest extends \PHPUnit_Framework_TestCase
      * @var Http\FileResolver
      */
     protected $_resolver;
+    // @codingStandardsIgnoreEnd
 
     /**
      * Sets the paths to files used in this test, and creates a shared resolver instance
@@ -76,7 +79,8 @@ class FileResolverTest extends \PHPUnit_Framework_TestCase
      */
     public function testSetFileInvalid()
     {
-        $this->setExpectedException('Zend\\Authentication\\Adapter\\Http\\Exception\\ExceptionInterface', 'Path not readable');
+        $this->expectException(Http\Exception\ExceptionInterface::class);
+        $this->expectExceptionMessage('Path not readable');
         $this->_resolver->setFile($this->_badPath);
     }
 
@@ -98,7 +102,8 @@ class FileResolverTest extends \PHPUnit_Framework_TestCase
      */
     public function testConstructInvalid()
     {
-        $this->setExpectedException('Zend\\Authentication\\Adapter\\Http\\Exception\\ExceptionInterface', 'Path not readable');
+        $this->expectException(Http\Exception\ExceptionInterface::class);
+        $this->expectExceptionMessage('Path not readable');
         $v = new Http\FileResolver($this->_badPath);
     }
 
@@ -109,7 +114,8 @@ class FileResolverTest extends \PHPUnit_Framework_TestCase
      */
     public function testResolveUsernameEmpty()
     {
-        $this->setExpectedException('Zend\\Authentication\\Adapter\\Http\\Exception\\ExceptionInterface', 'Username is required');
+        $this->expectException(Http\Exception\ExceptionInterface::class);
+        $this->expectExceptionMessage('Username is required');
         $this->_resolver->resolve('', '');
     }
 
@@ -120,7 +126,8 @@ class FileResolverTest extends \PHPUnit_Framework_TestCase
      */
     public function testResolveRealmEmpty()
     {
-        $this->setExpectedException('Zend\\Authentication\\Adapter\\Http\\Exception\\ExceptionInterface', 'Realm is required');
+        $this->expectException(Http\Exception\ExceptionInterface::class);
+        $this->expectExceptionMessage('Realm is required');
         $this->_resolver->resolve('username', '');
     }